Search-arrow-display-endpoint-data/requirements.txt
blinker==1.6.2
certifi==2023.5.7
chardet==4.0.0
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.9
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
requests==2.25.1
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.4
zipp==3.15.0